requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
idna==3.4
numpy==1.24.3
pandas==2.0.1
pycparser==2.21
PyGithub==1.58.1
PyJWT==2.6.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
six==1.16.0
tzdata==2023.3
urllib3==1.26.15
wrapt==1.15.0
